1) Renamed to have the postfix of `Dto` as this is a API data transfer object 2) Adds comments to indicate what request this is being used by i.e. POST, PUT, GET, etc
