--- data/src/main/java/me/androidbox/data/remote/model/AgendaDto.kt
@@ -1,9 +1,0 @@
-package me.androidbox.data.remote.model
-
-import me.androidbox.data.remote.model.request.EventRequestDto
-
-data class AgendaDto(
-    val listOfEvent: List<EventRequestDto>,
-    val listOfTask: List<TaskDto>,
-    val listOfReminder: List<ReminderDto>
-)
