--- data/src/main/java/me/androidbox/data/remote/model/ReminderDto.kt
@@ -1,9 +1,0 @@
-package me.androidbox.data.remote.model
-
-data class ReminderDto(
-    val id: String,
-    val title: String,
-    val description: String,
-    val time: Long,
-    val remindAt: Long
-)
