--- data/src/main/java/me/androidbox/data/remote/model/SyncAgendaDto.kt
@@ -1,7 +1,0 @@
-package me.androidbox.data.remote.model
-
-data class SyncAgendaDto(
-    val deletedEventId: List<String>,
-    val deletedTaskId: List<String>,
-    val deletedReminderId: List<String>
-)
