--- data/src/main/java/me/androidbox/data/remote/model/TaskDto.kt
@@ -1,10 +1,0 @@
-package me.androidbox.data.remote.model
-
-data class TaskDto(
-    val id: String,
-    val title: String,
-    val description: String,
-    val time: Long,
-    val remindAt: Long,
-    val isDone: Boolean
-)
