+++ data/src/main/java/me/androidbox/data/remote/model/request/EventRequestDto.kt
@@ -1,5 +1,15 @@
 package me.androidbox.data.remote.model.request
 
+
+/**
+ * /event
+ * POST/Multipart
+ *
+ * Creates a new event. Since Tasky supports offline mode, the IDs for agenda items
+ * need to be generated client-side. Please use UUID.randomUUID().toString() for that.
+ *
+ * Response body: EventDto
+ * */
 data class EventRequestDto(
     val id: String,
     val title: String,
@@ -7,5 +17,5 @@ data class EventRequestDto(
     val from: Long,
     val to: Long,
     val remindAt: Long,
-    val listOfAttendee: List<String>
+    val attendeeIds: List<String>
 )
