+++ data/src/main/java/me/androidbox/data/remote/model/request/EventUpdateRequestDto.kt
@@ -1,5 +1,16 @@
 package me.androidbox.data.remote.model.request
 
+
+/**
+ * /event
+ * PUT/Multipart
+ *
+ * Updates an existing event. In addition to creating an event, this also takes
+ * potentially deleted photos into consideration. Also, the isGoing boolean refers to the isGoing
+ * status of the local user.
+ *
+ * Response body: EventDto
+ * */
 data class EventUpdateRequestDto(
     val id: String,
     val title: String,
@@ -7,8 +18,8 @@ data class EventUpdateRequestDto(
     val from: Long,
     val to: Long,
     val remindAt: Long,
-    val listOfAttendee: List<String>, /* attendee id */
-    val listOfDeletedPhotoKey: List<String>,
+    val attendeeIds: List<String>, /* attendee id */
+    val deletedPhotoKeys: List<String>,
     val isGoing: Boolean
 )
