--- data/src/main/java/me/androidbox/data/remote/model/request/LoginResponseDto.kt
@@ -1,7 +1,0 @@
-package me.androidbox.data.remote.model.request
-
-data class LoginResponseDto(
-    val token: String,
-    val userId: String,
-    val fullName: String
-)
