+++ data/src/main/java/me/androidbox/data/remote/model/request/TaskRequestDto.kt
@@ -1,0 +1,22 @@
+package me.androidbox.data.remote.model.request
+
+
+/**
+ * /task
+ * POST
+ * Creates a new task
+ *
+ * PUT
+ * Updates an existing task
+ *
+ * No Response body
+ *
+ * */
+data class TaskRequestDto(
+    val id: String,
+    val title: String,
+    val description: String,
+    val time: Long,
+    val remindAt: Long,
+    val isDone: Boolean
+)
