+++ data/src/main/java/me/androidbox/data/remote/model/response/AttendeeGetResponseDto.kt
@@ -1,7 +1,5 @@
 package me.androidbox.data.remote.model.response
 
-import me.androidbox.data.remote.model.AttendeeDto
-
 data class AttendeeGetResponseDto(
     val isExistingUser: Boolean,
     val attendee: AttendeeDto
