+++ data/src/main/java/me/androidbox/data/remote/model/response/ErrorResponseDto.kt
@@ -1,5 +1,6 @@
 package me.androidbox.data.remote.model.response
 
+
 data class ErrorResponseDto(
     val message: String
 )
