+++ data/src/main/java/me/androidbox/data/remote/model/response/GetAttendeeDto.kt
@@ -1,0 +1,14 @@
+package me.androidbox.data.remote.model.response
+
+/**
+ * /attendee
+ * GET
+ * Checks if a user with a given email exists and can be added as attendee to an event.
+ *
+ * Query parameters: email
+ * Response: GetAttendeeDto
+ * */
+data class GetAttendeeDto(
+    val doesUserExist: Boolean,
+    val attendee: AttendeeDto
+)
