+++ data/src/main/java/me/androidbox/data/remote/model/response/LoginDto.kt
@@ -1,0 +1,7 @@
+package me.androidbox.data.remote.model.response
+
+data class LoginDto(
+    val token: String,
+    val userId: String,
+    val fullName: String
+)
