--- data/src/main/java/me/androidbox/data/remote/model/response/LoginRequestDto.kt
@@ -1,6 +1,0 @@
-package me.androidbox.data.remote.model.response
-
-data class LoginRequestDto(
-    val email: String,
-    val password: String
-)
