+++ data/src/main/java/me/androidbox/data/remote/model/response/PhotoDto.kt
@@ -1,4 +1,4 @@
-package me.androidbox.data.remote.model
+package me.androidbox.data.remote.model.response
 
 data class PhotoDto(
     val key: String,
